terms of use for images on dtek online store
Images used by DTEK Online Store are royalty free public domain photos from Pexels. Each of the photos have been altered by adding the word "DTEK" to the bottom right corner and can therefore be used and/or sold in their altered form by DTEK, in accordance with Pexels published terms of use as described in this document. The images may be sold by DTEK because DTEK Online Store is not a stock photo or wallpaper platform, the images have been altered, and no endorsement of DTEK products is implied on the imagery.